Replaced MSVC stuff with Makefile stuff
