+++ .gitignore
@@ -1,3 +1,2 @@
-Project1/x64
-Project1/Debug
-x64
+/bin
+.vscode
