+++ Makefile
@@ -1,0 +1,47 @@
+TARGET_EXEC ?= square
+
+CC=gcc
+CXX=g++
+
+BIN_DIR ?= ./bin
+SRC_DIRS ?= ./src
+
+LIBS := sfml-graphics sfml-window sfml-system
+SRCS := $(shell find $(SRC_DIRS) -name *.cpp -or -name *.c -or -name *.s)
+OBJS := $(SRCS:%=$(BIN_DIR)/%.o)
+DEPS := $(OBJS:.o=.d)
+
+INC_DIRS := $(shell find $(SRC_DIRS) -type d)
+INC_FLAGS := $(addprefix -I,$(INC_DIRS))
+
+LDFLAGS := $(shell pkg-config --libs $(LIBS))
+
+CPPFLAGS := $(INC_FLAGS) -g -MMD -MP $(shell pkg-config --cflags $(LIBS))
+
+$(BIN_DIR)/$(TARGET_EXEC): $(OBJS)
+	$(CXX) $(OBJS) -o $@ $(LDFLAGS)
+
+# assembly
+$(BIN_DIR)/%.s.o: %.s
+	$(MKDIR_P) $(dir $@)
+	$(AS) $(ASFLAGS) -c $< -o $@
+
+# c source
+$(BIN_DIR)/%.c.o: %.c
+	$(MKDIR_P) $(dir $@)
+	$(CC) $(CPPFLAGS) $(CFLAGS) -c $< -o $@
+
+# c++ source
+$(BIN_DIR)/%.cpp.o: %.cpp
+	$(MKDIR_P) $(dir $@)
+	$(CXX) $(CPPFLAGS) $(CXXFLAGS) -c $< -o $@
+
+
+.PHONY: clean
+
+clean:
+	$(RM) -r $(BIN_DIR)
+
+-include $(DEPS)
+
+MKDIR_P ?= mkdir -p
